progress

+++ dev/src/dataAccessLayer/transportModule/DistanceBetweenSites.java
@@ -1,6 +1,6 @@
 package dataAccessLayer.transportModule;
 
-public record DistanceBetweenSites(String source, String destination, int distance) {
+public record DistanceBetweenSites(String source, String destination, double distance) {
     public static DistanceBetweenSites getLookupObject(String source, String destination) {
         return new DistanceBetweenSites(source,destination, 0);
     }
